+++ src/app/core/auth.service.ts
@@ -645,7 +645,7 @@ export class AuthService {
           // used in signOut > removeInstanceId
           this.userId = user._id;
 
-          // console.log('[AUTH-SERV] isActivePAY in signin ', this.isActivePAY)
+          // this.logger.log('[AUTH-SERV] isActivePAY in signin ', this.isActivePAY)
           // this.sleekplanSso(user, this.isActivePAY)
 
         }
@@ -1047,29 +1047,29 @@ export class AuthService {
 
 
     // Force reload the page
-    // console.log('Reloading page to reset Sleekplan state.');
+    // this.logger.log('Reloading page to reset Sleekplan state.');
     // window.location.reload();
 
     // this.removeSleekScript();
     // this.sleekplanService.hasLogout()
     // if (window.hasOwnProperty('SLEEK_USER')) { 
-    //   console.log('[AUTH-SERV] SLEEK_USER window ', window['SLEEK_USER']);
+    //   this.logger.log('[AUTH-SERV] SLEEK_USER window ', window['SLEEK_USER']);
     //   delete window['SLEEK_USER'];
-    //   console.log('[AUTH-SERV] SLEEK_USER has been removed from the window object.');
+    //   this.logger.log('[AUTH-SERV] SLEEK_USER has been removed from the window object.');
     //   // this.deleteCookie('_sleek_product');
-    //   // console.log('Current cookies:', document.cookie);
+    //   // this.logger.log('Current cookies:', document.cookie);
     // } else {
-    //   console.warn('SLEEK_USER does not exist on the window object.');
+    //   this.logger.warn('SLEEK_USER does not exist on the window object.');
     // }
 
     // if (window.hasOwnProperty('SLEEK_PRODUCT_ID')) { 
-    //   console.log('[AUTH-SERV] SLEEK_PRODUCT_ID window ', window['SLEEK_PRODUCT_ID']);
+    //   this.logger.log('[AUTH-SERV] SLEEK_PRODUCT_ID window ', window['SLEEK_PRODUCT_ID']);
     //   delete window['SLEEK_PRODUCT_ID'];
-    //   console.log('[AUTH-SERV] SLEEK_PRODUCT_ID has been removed from the window object.');
+    //   this.logger.log('[AUTH-SERV] SLEEK_PRODUCT_ID has been removed from the window object.');
     //   // this.deleteCookie('_sleek_product');
-    //   // console.log('Current cookies:', document.cookie);
+    //   // this.logger.log('Current cookies:', document.cookie);
     // } else {
-    //   console.warn('SLEEK_PRODUCT_ID does not exist on the window object.');
+    //   this.logger.warn('SLEEK_PRODUCT_ID does not exist on the window object.');
     // }
   }
 
@@ -1221,6 +1221,7 @@ export class AuthService {
 
     this.user_bs.next(null)
     this.project_bs.next(null)
+    
     this.logger.log('[AUTH-SERV] SIGNOUT project_bs VALUE: ', this.project_bs.value)
 
     const storedRoute = this.localDbService.getFromStorage('wannago')
@@ -1232,6 +1233,8 @@ export class AuthService {
     localStorage.removeItem('user')
     localStorage.removeItem('project')
     localStorage.removeItem('role')
+    localStorage.removeItem('current_project_user')
+
 
     // if (calledby !== 'autologin') {
     this.logger.log('[AUTH-SERV] Signout this.router.url +++++ ', this.router.url)
@@ -1443,8 +1446,8 @@ export class AuthService {
 
   public siginWithGoogle() {
     // this.logger.log('[AUTH-SERV] siginWithGoogle HERE YES!!!')
-    // const url = this.SERVER_BASE_PATH + "auth/google"
-    const url = "https://eu.rtmv3.tiledesk.com/api/auth/google"
+    const url = this.SERVER_BASE_PATH + "auth/google"
+    // const url = "https://eu.rtmv3.tiledesk.com/api/auth/google"
     window.open(url, '_self');
 
     this.localDbService.setInStorage('swg', 'true')
@@ -1452,8 +1455,8 @@ export class AuthService {
   }
 
   public siginUpWithGoogle() {
-
-    const url = "https://eu.rtmv3.tiledesk.com/api/auth/google?redirect_url=%23%2Fcreate-project-gs"
+    const url = this.SERVER_BASE_PATH + "auth/google?redirect_url=%23%2Fcreate-project-gs"
+    // const url = "https://eu.rtmv3.tiledesk.com/api/auth/google?redirect_url=%23%2Fcreate-project-gs"
 
     // this.logger.log('siginUpWithGoogle ', url)
     window.open(url, '_self');
@@ -1462,7 +1465,7 @@ export class AuthService {
 
   signinWithOAuth2() {
     const url = this.SERVER_BASE_PATH + 'auth/oauth2'
-    console.log('[AUTH-SERV] signinWithOAuth2 url ', url)
+    this.logger.log('[AUTH-SERV] signinWithOAuth2 url ', url)
     window.open(url, '_self');
   }