+++ src/app/knowledge-bases/modals/modal-add-namespace/modal-add-namespace.component.ts
@@ -1,5 +1,6 @@
 import { Component, Inject, OnInit } from '@angular/core';
 import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
+import { BrandService } from 'app/services/brand.service';
 
 @Component({
   selector: 'appdashboard-modal-add-namespace',
@@ -7,17 +8,34 @@ import { MAT_DIALOG_DATA, MatDialogRef } from '@angular/material/dialog';
   styleUrls: ['./modal-add-namespace.component.scss']
 })
 export class ModalAddNamespaceComponent implements OnInit {
-  public namespaceName: string
+  public namespaceName: string;
+  selectedNamespaceType: string = "standard";
+  hybridActive: boolean = false;
+  salesEmail: string;
+
   constructor(
     @Inject(MAT_DIALOG_DATA) public data: any,
     public dialogRef: MatDialogRef<ModalAddNamespaceComponent>,
-  ) { }
+    public brandService: BrandService,
+  ) { 
+    this.hybridActive = data.hybridActive;
+
+    const brand = brandService.getBrand();
+    this.salesEmail = brand['CONTACT_SALES_EMAIL'];
+  }
 
   ngOnInit(): void {
   }
+  
+  contactUs() {
+    window.open(`mailto:${this.salesEmail}?subject=Increase quotas`);
+  }
 
-  onOkPresssed(namespaceName ){
-    this.dialogRef.close({'namespaceName': namespaceName });
+  onOkPresssed(namespaceName, selectedNamespaceType){
+    this.dialogRef.close({
+      'namespaceName': namespaceName,
+      'hybrid': selectedNamespaceType !== 'standard'
+    });
   }
 
   onNoClick(): void {