+++ src/app/users/users.component.ts
@@ -47,11 +47,13 @@ export class UsersComponent extends PricingBaseComponent implements OnInit, Afte
   id_projectUser: string
   user_firstname: string
   user_lastname: string
+  user_fullname: string
   user_id: string
 
   // set to none the property display of the modal
   display = 'none'
   displayCancelInvitationModal = 'none'
+  displayRestoreModal = 'none';
   project: Project
 
   USER_ROLE: string
@@ -950,7 +952,26 @@ export class UsersComponent extends PricingBaseComponent implements OnInit, Afte
     this.user_firstname = userFirstname
     this.user_lastname = userLastname
 
+    this.user_fullname = ''
+    if (this.user_firstname && this.user_lastname) {
+        this.user_fullname = this.user_firstname + ' ' + this.user_lastname
+    }
+    if (this.user_firstname && !this.user_lastname) {
+        this.user_fullname = this.user_firstname
+    }
+     if (!this.user_firstname && this.user_lastname) {
+        this.user_fullname = this.user_lastname
+    }
+
     this.logger.log('[USERS] OPEN DELETE MODAL - PROJECT-USER with ID ',  this.id_projectUser, ' - (Firstname: ',  userFirstname, '; Lastname: ',    userLastname,  ')')
+  }
+
+  openRestoreModal(projectUser_id: string, userID: string, userFirstname: string, userLastname: string) {
+    this.displayRestoreModal = 'block';
+    this.id_projectUser = projectUser_id
+    this.user_id = userID
+    this.user_firstname = userFirstname
+    this.user_lastname = userLastname
   }
 
   onCloseDeleteModalHandled() {
@@ -1023,19 +1044,33 @@ export class UsersComponent extends PricingBaseComponent implements OnInit, Afte
         // NOTIFY SUCCESS
         this.notify.showWidgetStyleUpdateNotification(this.deleteProjectUserSuccessNoticationMsg, 2, 'done')
 
-        for (let i = 0; i < this.projectUsersList.length; i++) {
-          if (this.id_projectUser === this.projectUsersList[i]._id) {
-            this.projectUsersList.splice(i, 1)
-            this.projectUsersLength = this.projectUsersList.length
-            localStorage.removeItem('dshbrd----' + this.id_projectUser)
-          }
+        const userToDisable = this.projectUsersList.find(user => user._id === this.id_projectUser);
+        if (userToDisable) {
+          userToDisable.status = 'disabled';
         }
       },
     )
   }
 
+  onCloseRestoreModalHandled() {
+    this.displayRestoreModal = 'none';
+    this.usersService.updateProjectUser(this.id_projectUser, false, "", 'active').subscribe((projectUser: any) => {
+
+    }, (error) => {
+      this.logger.error('[USERS] ON-CLOSE-RESTORE-MODAL - RESTORE PROJECT USERS - ERROR ', error)
+      this.notify.showWidgetStyleUpdateNotification("An error occurred restoring teammate", 4, 'report_problem')
+    }, () => {
+      this.notify.showWidgetStyleUpdateNotification("Teammate successfully restored", 2, 'done')
+      const userToEnable = this.projectUsersList.find(user => user._id === this.id_projectUser);
+      if (userToEnable) {
+        userToEnable.status = 'active';
+      }
+    })
+  }
+
   onCloseModal() {
-    this.display = 'none'
+    this.display = 'none';
+    this.displayRestoreModal = 'none';
   }